rename upgrade and upgrade task packages
